% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_measures_summary}
\alias{create_measures_summary}
\title{Creates a measures summary from a Power BI report list}
\usage{
create_measures_summary(report)
}
\arguments{
\item{report}{A Power BI report list read in using read_pbi or read_layout}
}
\value{
tibble
}
\description{
Creates a measures summary from a Power BI report list
}
\examples{
\dontrun{
create_measures_summary(report)
}
}
\keyword{extract}
\keyword{measures,}
