% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_qa_report}
\alias{create_qa_report}
\title{Create quality assurance report from Power BI list}
\usage{
create_qa_report(report, file = "pbi_qa_report.xlsx")
}
\arguments{
\item{report}{A Power BI report list read in using read_pbi or read_layout}
}
\value{
xlsx
}
\description{
Create quality assurance report from Power BI list
}
\examples{
\dontrun{
create_qa_report(report)
}
}
\keyword{qa}
\keyword{report,}
