% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_report_summary}
\alias{create_report_summary}
\title{Provides a high level summary of a report}
\usage{
create_report_summary(report)
}
\arguments{
\item{report}{A Power BI report list read in using read_pbi or read_layout}
}
\value{
tibble
}
\description{
Provides a high level summary of a report
}
\examples{
\dontrun{
create_report_summary(report)
}
}
\keyword{extract}
\keyword{filters,}
\keyword{report,}
