% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_section_summary}
\alias{create_section_summary}
\title{Provides a high level summary of a section}
\usage{
create_section_summary(section)
}
\arguments{
\item{section}{A section element in a Power BI list}
}
\value{
tibble
}
\description{
Provides a high level summary of a section
}
\examples{
\dontrun{
pbiAssure:::create_section_summary(report$sections[[6]])
}
}
\keyword{extract}
\keyword{filters,}
\keyword{section,}
