% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{create_sections_summary}
\alias{create_sections_summary}
\title{Provides a high level summary of all report sections}
\usage{
create_sections_summary(report)
}
\arguments{
\item{report}{A Power BI report list read in using read_pbi or read_layout}
}
\value{
tibble
}
\description{
Provides a high level summary of all report sections
}
\examples{
\dontrun{
create_sections_summary(report)
}
}
\keyword{extract}
\keyword{filters,}
\keyword{section,}
