% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_entity_measures}
\alias{extract_entity_measures}
\title{Extract an individual measure from an individual entity (table) within a power bi report list.}
\usage{
extract_entity_measures(entity)
}
\value{
tibble
}
\description{
Extract an individual measure from an individual entity (table) within a power bi report list.
}
\examples{
\dontrun{
pbiAssure:::extract_entity_measures(RJSONIO::fromJSON(report$config)$modelExtensions[[1]]$entities[[1]])
}
}
\keyword{extract}
\keyword{measures,}
