% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_extensions_measures}
\alias{extract_extensions_measures}
\title{Extract all measures associated with all model extensions.}
\usage{
extract_extensions_measures(modelExtensions)
}
\value{
tibble
}
\description{
Extract all measures associated with all model extensions.
}
\examples{
\dontrun{
pbiAssure:::extract_extensions_measures(RJSONIO::fromJSON(report$config)$modelExtensions[[1]])
}
}
\keyword{extract}
\keyword{measures,}
