% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_filter}
\alias{extract_filter}
\title{Extract invividual filter from filters element of Power BI list}
\usage{
extract_filter(filters_element)
}
\arguments{
\item{filters_element}{A filters element in Power BI list}
}
\value{
tibble
}
\description{
Extract invividual filter from filters element of Power BI list
}
\examples{
\dontrun{
pbiAssure:::extract_filter(RJSONIO::fromJSON(report$sections[[6]]$visualContainers[[4]]$filters)[[1]])
}
}
\keyword{extract}
\keyword{filters,}
