% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_filter_where_in_value}
\alias{extract_filter_where_in_value}
\title{Extract value from where in filter in Power BI list}
\usage{
extract_filter_where_in_value(where)
}
\arguments{
\item{where_element}{A where in element in Power BI list}
}
\value{
tibble
}
\description{
Extract value from where in filter in Power BI list
}
\examples{
\dontrun{
pbiAssure:::extract_filter_where_in_value(RJSONIO::fromJSON(report$sections[[6]]$visualContainers[[4]]$filters)[[2]]$filter$Where[[1]])
}
}
\keyword{extract}
\keyword{filters,}
