% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_filters}
\alias{extract_filters}
\title{Extract all filters from an object}
\usage{
extract_filters(object)
}
\arguments{
\item{object}{A Power BI object that has attached filters}
}
\value{
tibble
}
\description{
Extract all filters from an object
}
\examples{
\dontrun{
pbiAssure:::extract_filters(report$sections[[6]]$visualContainers[[4]])
}
}
\keyword{extract}
\keyword{filters,}
