% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_measure}
\alias{extract_measure}
\title{Extract details of an individual power bi measure}
\usage{
extract_measure(measure)
}
\value{
tibble
}
\description{
Extract details of an individual power bi measure
}
\examples{
\dontrun{
pbiAssure:::extract_measure(RJSONIO::fromJSON(report$config)$modelExtensions[[1]]$entities[[1]]$measures[[1]])de
}
}
\keyword{extract}
\keyword{measure,}
