% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_section_visuals}
\alias{extract_section_visuals}
\title{Create a summmary of all Power BI visuals in a section}
\usage{
extract_section_visuals(section)
}
\arguments{
\item{section}{A section element in a Power BI list}
}
\value{
tibble
}
\description{
Create a summmary of all Power BI visuals in a section
}
\examples{
\dontrun{
pbiAssure:::extract_section_visuals(report$sections[[6]])
}
}
\keyword{extract}
\keyword{filters,}
\keyword{visual,}
