% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_single_visual}
\alias{extract_single_visual}
\title{Create a summmary of a single Power BI visual}
\usage{
extract_single_visual(single_visual)
}
\arguments{
\item{single_visual}{A visual container element in a Power BI list}
}
\value{
tibble
}
\description{
Create a summmary of a single Power BI visual
}
\examples{
\dontrun{
pbiAssure:::extract_single_visual(report$sections[[6]]$visualContainers[[2]])
}
}
\keyword{extract}
\keyword{filters,}
\keyword{visual,}
