% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_textbox}
\alias{extract_textbox}
\title{Extract text from a text box visual}
\usage{
extract_textbox(single_visual)
}
\arguments{
\item{single_visual}{A visual container element in a Power BI list}
}
\value{
tibble
}
\description{
Extract text from a text box visual
}
\examples{
\dontrun{
pbiAssure:::extract_single_visual(report$sections[[3]]$visualContainers[[2]])
}
}
\keyword{extract}
\keyword{text,}
\keyword{visual,}
