% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{if_null_na}
\alias{if_null_na}
\title{Helper function hat converts nulls to NA.}
\usage{
if_null_na(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
character
}
\description{
Helper function hat converts nulls to NA.
}
\examples{
\dontrun{
if_null_na(x)
}
}
\keyword{helper,}
\keyword{na}
\keyword{null,}
