% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_layout}
\alias{read_layout}
\title{Read a Power BI layout file into R as a list.}
\usage{
read_layout(file)
}
\arguments{
\item{file}{Power BI layout file}
}
\value{
list
}
\description{
Read a Power BI layout file into R as a list.
}
\keyword{layout}
\keyword{read,}
