% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_pbi}
\alias{read_pbi}
\title{Read a Power BI file into R as a list.}
\usage{
read_pbi(file)
}
\arguments{
\item{file}{Power BI file}
}
\value{
list
}
\description{
Read a Power BI file into R as a list.
}
\examples{
\dontrun{
read_layout("example.pbix")
}
}
\keyword{layout}
\keyword{read,}
